https://github.com/umjammer/vavi-sound-dx7/security/code-scanning/1?query=ref%3Arefs%2Fheads%2Fmain

+++ src/main/java/vavi/sound/dx7/FmOpKernel.java
@@ -27,7 +27,7 @@ class FmOpKernel {
             for (int i = 0; i < Note.N; i++) {
                 gain += dgain;
                 int y = Sin.lookup(phase + input[i]);
-                output[i] += ((long) y * (long) gain) >> 24;
+                output[i] += (int) (((long) y * (long) gain) >> 24);
                 phase += freq;
             }
         } else {
